# doc/api.yaml
openapi: 3.0.1
info:
  title: WASAText API
  version: 1.0.0
  description: API per l'app di messaggistica WASAText
servers:
  - url: http://0.0.0.0:3000
...
security:
  - bearerAuth: []
paths:
  /session:
    post:
      tags: 
        - auth
      summary: Login o registrazione utente
      description: Esegue il login o registra un nuovo utente se non esiste già nel sistema.
      operationId: doLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login avvenuto con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '201':
          description: Registrazione avvenuta con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /conversations/{id}/messages:
    post:
      tags:
        - message
      summary: Invia un messaggio in una conversazione
      description: L’utente autenticato invia un nuovo messaggio (testuale o con media) in una conversazione 1:1 o di gruppo a cui partecipa.
      operationId: sendMessage
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Messaggio inviato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /messages/{id}:
    delete:
      tags:
        - message
      summary: Elimina un messaggio inviato
      description: Elimina un messaggio inviato dall’utente autenticato. L’operazione è consentita solo per messaggi creati da lui.
      operationId: deleteMessage
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Messaggio eliminato con successo
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /messages/{id}/forward:
    post:
      tags:
        - message
      summary: Inoltra un messaggio a un'altra conversazione
      description: Permette all’utente autenticato di inoltrare un messaggio già inviato in una nuova conversazione (1:1 o di gruppo).
      operationId: forwardMessage
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idConversation
              properties:
                idConversation:
                  $ref: '#/components/schemas/id'
      responses:
        '201':
          description: Messaggio inoltrato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /messages/{id}/reactions:
    post:
      tags:
        - reaction
      summary: Aggiunge una reazione a un messaggio
      description: Permette all’utente autenticato di aggiungere una reazione (emoji) a un messaggio, se è membro della conversazione a cui appartiene il messaggio.
      operationId: commentMessage
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactToMessageRequest'
      responses:
        '201':
          description: Reazione aggiunta con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /messages/{id}/reactions/me:
    delete:
      tags:
        - reaction
      summary: Rimuove la reazione dell’utente autenticato
      description: Rimuove la reazione precedentemente aggiunta al messaggio da parte dell’utente autenticato. Funziona solo se l’utente aveva già reagito al messaggio.
      operationId: uncommentMessage
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Reazione rimossa con successo
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /messages/{id}/status:
    put:
      tags:
        - status
      summary: Aggiorna lo stato del messaggio per l’utente autenticato
      description: Imposta i campi `delivered` e/o `seen` a `true` per il messaggio specificato, per l’utente autenticato. Solo `true` è permesso (non si può annullare).
      operationId: updateMessageStatus
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMessageStatusRequest'
      responses:
        '204':
          description: Stato aggiornato con successo. Nessun contenuto restituito.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Dati non validi
    Unauthorized:
      description: Accesso negato (token mancante o non valido)
    Forbidden:
      description: Accesso negato (permessi insufficienti)
    NotFound:
      description: Risorsa non trovata
...
    InternalServerError:
      description: Errore interno del server
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/id'
      description: ID numerico della risorsa nella path
  schemas:
...
    id:
      type: integer
      format: int32
      description: ID univoco intero
      minimum: 1
      example: 389
    User:
      type: object
      required:
        - uuid
        - username
      description: Identifica un utente
      properties:
        uuid:
          type: string
          format: uuid
          example: 6e9f8a42-1234-5678-90ab-cdef12345678
          description: identificatore univoco utente
        username:
          type: string
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_]+$'
          description: Username dell'utente
        photoUrl:
          type: string
          pattern: '^https?://.*'
          example: 'https://www.example.com/imgs/photo.png'
          description: URL della foto profilo dell’utente
    Message:
      type: object
      required:
        - id
        - type
        - content
        - mediaUrl
        - timestamp
        - idConversation
        - uuidSender
      description: Identifica un messaggio
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          example: 389
          description: ID univoco di un messaggio
        type:
          type: string
          enum:
          - text
          - photo
          description: indica il tipo di messaggio inviato 'text' per solo testuale 'photo' per l'invio di una foto con anche il testo (opzionale)
          example: text
        content:
          type: string
          minLength: 0
          maxLength: 500
          example: "Ciao a tutti! 😀"
          description: Stringa contenente il contenuto testuale del messaggio
          nullable: true
        mediaUrl:
          type: string
          format: binary
          nullable: true
          description: Foto da inviare (può essere null)
        timestamp:
          type: string
          format: date-time
          example: '2025-05-30T14:45:00Z'
          description: Data e ora di invio del messaggio
        idRepliesTo:
          type: integer
          format: int32
          example: 389
          nullable: true
          description: ID univoco del messaggio a cui si riferisce in caso di risposta ad un messaggio
        idConversation:
          type: integer
          format: int32
          minimum: 1
          example: 389
          description: ID della conversazione in cui il messaggio è stato mandato
        uuidSender:
          type: string
          format: uuid
          example: 6e9f8a42-1234-5678-90ab-cdef12345678
          description: identificatore univoco utente che ha inviato il messaggio
        idForwardedFrom:
          type: integer
          format: int32
          example: 123
          nullable: true
          description: ID del messaggio originale se il messaggio è inoltrato
    Reaction:
      type: object
      required:
        - uuidUser
        - idMessage
        - emoji
      description: Identifica una reazione ad un messaggio fatta da un determinato utente
      properties:
        uuidUser:
          type: string
          format: uuid
          example: 6e9f8a42-1234-5678-90ab-cdef12345678
          description: UUID dell’utente che ha inviato la reazione
        idMessage:
          type: integer
          format: int32
          minimum: 1
          example: 389
          description: ID del messaggio a cui è stata applicata la reazione
        emoji:
          type: string
          description: >- 
            Emoji Unicode singola (attenzione: alcune emoji possono essere composte da più codepoint).
          example: "🔥"
          minLength: 1
          maxLength: 8
...
    LoginRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_]+$'
          description: Username scelto dall’utente
...
    SendMessageRequest:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - text
            - photo
          description: Tipo di messaggio
          example: 'text'
        content:
          type: string
          maxLength: 500
          example: "Ciao a tutti! 😀"
          description: Stringa contenente il contenuto testuale del messaggio
          nullable: true
        mediaUrl:
          type: string
          format: url
          nullable: true
          example: 'https://www.example.com/imgs/photo.png'
          description: URL della foto da inviare (può essere null)
        idRepliesTo:
          type: integer
          format: int32
          example: 389
          nullable: true
          description: ID univoco del messaggio a cui si riferisce in caso di risposta ad un messaggio
    ReactToMessageRequest:
      type: object
      required:
        - emoji
      properties:
        emoji:
          type: string
          description: >- 
           Emoji Unicode singola (attenzione: alcune emoji possono essere composte da più codepoint).
          example: "🔥"
          minLength: 1
          maxLength: 8
    UpdateMessageStatusRequest:
      type: object
      required:
        - delivered
        - seen
      properties:
        delivered:
          type: boolean
          description: Consegnato true o false
          example: true
        seen:
          type: boolean
          description: visualizzato true o false
          example: true